code cleanup and conflict resolved
